In 2.x/3.x, this file was lib/libmsg/addr.c.
In 4.x, it was lib/libmsg/addrutil.cpp.  (They felt the need to convert it
to C++ for no adequately explainable reason, and to add some dependencies
on the rest of libmsg.)

Since libmime needs this, and libmsg isn't being built, I converted it
back to C, removed the libmsg dep, and put a copy of this file here.

Someday, something more sensible should be done.  Like deleting the copy
in libmsg, perhaps.