bug 1520 (www.mirabilis.com) the problem was the way table layout apportioned height from a rowspanning cell. Now, we do 2 things correctly: 1) we handle rows with rowspanning cells and no other cells. 2) we divide the height of a rowspanning cell proportionately to the rows spanned, where the proportion is defined by the max height of a cell in that row that has no effective rowspan.
the partial fix for bug 1431 has to do with dividing width between percent-width and fixed-width columns in the same
table. You won't see any better layout for this page (ww.boston.com) as a result of the code I added.  The rest of this
fix is a bit risky, so I'm waiting until after stability push.