this call to GetFileLocation() is not needed.
we were calling it so that as a side effect, it would set the global profile
and prefs dir, but we don't need to do this.  We are already calling it
directly before we get here.

other changes to move #ifdef to the first column (follow convention) and to
get rid of four warings,