Support for ncols and installcmap prefs.
Currently, 8 bit pseudocolor visual users are going to get a private cmap
installed base on design of gdk_rgb* code design. We want the default to
really be that we share with the server default colormap; if that looks
too bad (e.g., can't get enough colors, user can set the browser.installcmap
pref to "true" to get a priviate cmap installed.

Gdk determines if to install a private cmap based on being able to allocate
a certain number of colors from the default colormap. If that fails, a
private cmap is installed. The default gdk threshold is 5 * 5 * 5. If we
see browser.installcmap == true, we set to 255 to force it. If we see
browser.installcmap == false, we set it to 0. Also added browser.ncols
pref, which allows fine-grained control of this; the user can set the
threshold value and control when gdk will use shared colormap vs. private.