If we make an error string have a negative ID then the high bit of the id is set. This allows us to pass in the string ID around as an nsresult and it will properly be detected as an error code when we do NS_FAILED(rv) tests on it..nifty huh...
In order to do this, we need to get rid of the enum for the string ids. This also makes things more robust
as we must now explicitly list the id for each string id instead of using the enum to do it for us. It was easy
to get out of step with the properties file when the enum was used.