r=fur@netscape.com, approval=chofmann@netscape.com.  Fix for bug 5369 ("[PP]Crash on startup because of Java Plug-in 1.3 for Netscape Navigator.")  Make the code more bulletproof by checking for NULL arguments to several methods.  If JVM isn't running, jEnv arg should be NULL.
This doesn't entirely fix the problem of a JVM gone awry in all cases;
in the case of this particular bug, the Sun JDK 1.3 Beta HotSpot VM
caused an error, and there's no way for Mozilla code to detect this
failure mode.  The fix for that half of the problem is to use the new
JDK 1.3 Early Access VM, which does not exhibit the problem.