Bug #37058
Added a simple test to CreateInstance, similar to the existing test in
GetService in the service manager, to prevent instances from being created
during shutdown.  We detected no calls to CreateInstance in normal code we
tested during shutdown.  If such occur, the fix is NOT to back out the
check, but rather to eliminate the calls to CreateInstance either by
registering a shutdown observer which gets called just before the
shutdown, or creating the instance before shutdown.

r=scc