stop the crash that causes #40306.  the "first time start page" causes us many problem on linux, not sure if it is related to that page, or the redirect.  still investigating.
other changes here remove the crap code that forces you to go to
the checkin guidelines page and to see the account manager on start up
on debug builds.  this was preventing debug builds from seeing the problem.

after this change debug behaves like optimized, which is what we want.

r=mscott, a=granrose